+++ cppcache/include/geode/UserFunctionExecutionException.hpp
@@ -81,7 +81,7 @@ class UserFunctionExecutionException : public Serializable {
    * Note that you must implement this only if you use the HeapLRU feature.
    * @throws IllegalStateException If this api is called from User code.
    */
-  virtual uint32_t objectSize() const override;
+  virtual size_t objectSize() const override;
 
   /**
    *@brief return the typeId byte of the instance being serialized.
