+++ cppcache/src/GetAllServersRequest.hpp
@@ -47,7 +47,7 @@ class GetAllServersRequest : public Serializable {
   virtual int8_t DSFID() const {
     return static_cast<int8_t>(GeodeTypeIdsImpl::FixedIDByte);
   }
-  virtual uint32_t objectSize() const { return m_serverGroup->length(); }
+  virtual size_t objectSize() const { return m_serverGroup->length(); }
   virtual ~GetAllServersRequest() {}
 };
