+++ cppcache/src/GetAllServersResponse.hpp
@@ -45,9 +45,7 @@ class GetAllServersResponse : public Serializable {
   virtual int8_t DSFID() const {
     return static_cast<int8_t>(GeodeTypeIdsImpl::FixedIDByte);
   }
-  virtual uint32_t objectSize() const {
-    return static_cast<uint32_t>(m_servers.size());
-  }
+  virtual size_t objectSize() const { return m_servers.size(); }
   std::vector<ServerLocation> getServers() { return m_servers; }
   virtual ~GetAllServersResponse() {}
 };
