+++ cppcache/src/LocatorListResponse.cpp
@@ -30,7 +30,7 @@ int8_t LocatorListResponse::typeId() const {
   return static_cast<int8_t>(GeodeTypeIdsImpl::LocatorListResponse);
 }
 
-uint32_t LocatorListResponse::objectSize() const { return 0; }
+size_t LocatorListResponse::objectSize() const { return 0; }
 
 void LocatorListResponse::readList(DataInput& input) {
   uint32_t size = input.readInt32();
