+++ cppcache/src/ServerLocationRequest.hpp
@@ -33,7 +33,7 @@ class ServerLocationRequest : public Serializable {
   virtual int32_t classId() const;
   virtual int8_t typeId() const = 0;
   virtual int8_t DSFID() const;
-  virtual uint32_t objectSize() const = 0;
+  virtual size_t objectSize() const = 0;
   virtual ~ServerLocationRequest() {}
 };
